ismn5650-lrd0036/Assignment7/requirements.txt
flask==3.1.2
python-dotenv==1.0.1
requests==2.31.0
openai==1.12.0
